+++ app/src/main/java/org/simple/clinic/bp/assignbppassport/BpPassportUiActions.kt
@@ -1,0 +1,8 @@
+package org.simple.clinic.bp.assignbppassport
+
+import org.simple.clinic.facility.Facility
+
+interface BpPassportUiActions {
+  fun openPatientEntryScreen(facility: Facility)
+  fun closeSheet()
+}